requirements.txt
Django
future
six
requests
#pytest
pytest-django
pytest-xdist
#coverage
openpyxl
lxml
futures
pytz
# TODO add wikiwho_simple as a package

djangorestframework
#coreapi
markdown
#django-filter
#django-crispy-forms
#django-guardian
django-rest-swagger
#drfdocs
python-memcached
drf-extensions

#django-admin-tools
#django-admin-tools-stats

django-extensions
psycopg2

django-registration
django-crispy-forms

#mediawiki-utilities
mwxml
mwreverts

# django-postgres-copy
